Moved evaluation to a separate module.
